fix!: Use double instead of float in map API (#4567)
* fix!: Use double instead of float in map API

Java numbers like 1.2 are by default doubles and having floats in the API only makes it harder to use

* adapt remaining code to use double

+++ vaadin-map-flow-parent/vaadin-map-flow/src/main/java/com/vaadin/flow/component/map/Map.java
@@ -299,7 +299,7 @@ public class Map extends MapBase {
      *
      * @return current zoom level
      */
-    public float getZoom() {
+    public double getZoom() {
         return getView().getZoom();
     }
 
@@ -318,7 +318,7 @@ public class Map extends MapBase {
      * @param zoom
      *            new zoom level
      */
-    public void setZoom(float zoom) {
+    public void setZoom(double zoom) {
         getView().setZoom(zoom);
     }
 }
